requirements.txt
flask==1.1.2
Pillow==7.2.0